requirements.txt
asgiref==3.5.2
certifi==2022.9.24
charset-normalizer==2.1.1
dj-database-url==1.0.0
Django==3.2.9
django-decouple==2.1
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
docker==6.0.0
idna==3.4
importlib-metadata==5.0.0
Markdown==3.4.1
packaging==21.3
Pillow==9.2.0
psycopg2-binary==2.9.5
PyJWT==2.6.0
pyparsing==3.0.9
python-decouple==3.6
pytz==2022.5
pywin32==304
requests==2.28.1
sqlparse==0.4.3
urllib3==1.26.12
websocket-client==1.4.1
whitenoise==6.2.0
zipp==3.10.0
